+++ src/scalars/schemaversion.ts
@@ -10,18 +10,18 @@ export const SchemaVersionScalar = new GraphQLScalarType({
   description: 'Date and Time in format yLLddHHmmss (20230101123456 for January 1, 2023 at 12:34:56). JSON parser should convert to javascript Date type.',
 
   serialize (value: unknown) {
-    return (value as DateTime).toFormat(schemaVersionFormat)
+    return (value as DateTime).toUTC().toFormat(schemaVersionFormat)
   },
 
   parseValue (value: unknown) {
-    const parsedValue = DateTime.fromFormat(value as string, schemaVersionFormat)
+    const parsedValue = DateTime.fromFormat(value as string, schemaVersionFormat, { zone: 'utc' })
     if (!parsedValue.isValid) throw new Error(`Invalid date: ${parsedValue.invalidReason!} ${parsedValue.invalidExplanation!}`)
     return parsedValue
   },
 
   parseLiteral (ast: ValueNode) {
     if (ast.kind === Kind.STRING) {
-      const parsedValue = DateTime.fromFormat(ast.value, schemaVersionFormat)
+      const parsedValue = DateTime.fromFormat(ast.value, schemaVersionFormat, { zone: 'utc' })
       if (!parsedValue.isValid) throw new Error(`Invalid date: ${parsedValue.invalidReason!} ${parsedValue.invalidExplanation!}`)
       return parsedValue
     }